Deploy execution files on cluster setup

+++ create_cluster.sh
@@ -15,6 +15,7 @@ set -u # Exit when undefined variable
 #
 
 SCRIPT_DIR=$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )
+MN_USER=bsc19533 # Also requires a MN_USER priv/public key with the same name
 
 
 #
@@ -79,6 +80,7 @@ setup_cluster() {
   echo "[INFO] Setting up cluster with ${NUM_NODES} nodes"
 
   # Retrieve master IP
+  local master_ip
   master_ip=$("${SCRIPT_DIR}"/create_cluster/get_node_ip.sh "${internal_props_file}" 0)
   echo "[INFO] MASTER NODE WILL RUN IN ${master_ip}"
 
@@ -94,12 +96,51 @@ setup_cluster() {
   scp "${SCRIPT_DIR}"/create_cluster/setup_cluster.sh "${USERNAME}"@"${master_ip}":.
   # shellcheck disable=SC2086  # Array split on purpose
   ssh "${USERNAME}"@"${master_ip}" ./setup_cluster.sh "${NODE_CPUS}" "${num_workers}" ${worker_ips[*]}
-  global_ev=$?
+  ev=$?
+  if [ "$ev" -ne 0 ]; then
+    echo "[ERROR] Cannot setup cluster"
+    exit $ev
+  fi
   ssh "${USERNAME}"@"${master_ip}" rm -f setup_cluster.sh
 
   echo "[INFO] Setting up cluster DONE"
+  global_ev=0
 }
 
+deploy_files() {
+  # Retrieve master IP
+  local master_ip
+  master_ip=$("${SCRIPT_DIR}"/create_cluster/get_node_ip.sh "${internal_props_file}" 0)
+
+  # Deploy scripts and input data
+  echo "[INFO] Deploying execution scripts to ${master_ip}..."
+  scp -v -r "${SCRIPT_DIR}"/execution/* "${USERNAME}@${master_ip}:/home/${USERNAME}/${BUCKET_NAME}/"
+  ev=$?
+  if [ "$ev" -ne 0 ]; then
+    echo "[ERROR] Cannot deploy execution scripts"
+    exit $ev
+  fi
+
+  echo "[INFO] Deploying SSH keys to ${master_ip}..."
+  scp ~/.ssh/${MN_USER}* "${USERNAME}@${master_ip}:/home/${USERNAME}/.ssh/"
+  ev=$?
+  if [ "$ev" -ne 0 ]; then
+    echo "[ERROR] Cannot deploy MN keys"
+    exit $ev
+  fi
+
+  echo "[INFO] Deploying input data to ${master_ip}..."
+  # shellcheck disable=SC2029  # We want variables to be expanded in client side
+  ssh "${USERNAME}@${master_ip}" scp -v -r -i "/home/${USERNAME}/.ssh/${MN_USER}" "${MN_USER}"@mn1.bsc.es:/gpfs/projects/bsc19/GUIDANCE/inputs "/home/${USERNAME}/${BUCKET_NAME}/"
+  ev=$?
+  if [ "$ev" -ne 0 ]; then
+    echo "[ERROR] Cannot deploy input data"
+    exit $ev
+  fi
+
+  echo "[INFO] Deploying files DONE"
+  global_ev=0
+}
 
 #
 # MAIN METHOD
@@ -146,6 +187,7 @@ main() {
   # Setup cluster
   if [ "${global_ev}" -eq 0 ]; then
     setup_cluster
+    deploy_files
   fi
 
   exit "${global_ev}"
